Add controller to add ownerReference to PVC
