+++ operator/charts/kit-operator/templates/controller/rbac.yaml
@@ -36,7 +36,7 @@ rules:
   resources: [leases]
   verbs: [create, get, patch, update, watch]
 - apiGroups: [""]
-  resources: [configmaps, nodes, secrets, services, statefulsets]
+  resources: [configmaps, nodes, secrets, services, statefulsets, persistentvolumeclaims]
   verbs: [create, get, list, patch, update, watch]
 - apiGroups: [apps]
   resources: [statefulsets, deployments, daemonsets]
